+++ src/modules/stereoDisparity/disparityThread.cpp
@@ -174,7 +174,7 @@ bool disparityThread::threadInit()
 
     Bottle p;
     igaze->getInfo(p);
-    int vHead=p.check(("head_version"),Value(1)).asInt();
+    int vHead=(int)p.check(("head_version"),Value(1.0)).asDouble();
 
     stringstream headType;
     headType << "v";
